+++ src/main/scala/com/github/unchama/seichiassist/util/Util.scala
@@ -1,6 +1,5 @@
 package com.github.unchama.seichiassist.util
 
-import org.bukkit._
 import org.bukkit.entity.LivingEntity
 import org.bukkit.event.entity.EntityDamageEvent.DamageCause
 
@@ -11,17 +10,6 @@ object Util {
   def isVotingFairyPeriod(start: Calendar, end: Calendar): Boolean = {
     val cur = Calendar.getInstance()
     cur.after(start) && cur.before(end)
-  }
-
-  def setDifficulty(worldNameList: List[String], difficulty: Difficulty): Unit = {
-    worldNameList.foreach { name =>
-      val world = Bukkit.getWorld(name)
-
-      if (world == null)
-        Bukkit.getLogger.warning(name + "という名前のワールドは存在しません。")
-      else
-        world.setDifficulty(difficulty)
-    }
   }
 
   /**
