+++ src/main/scala/com/github/unchama/seichiassist/util/WorldSetting.scala
@@ -1,0 +1,16 @@
+package com.github.unchama.seichiassist.util
+
+import org.bukkit.{Bukkit, Difficulty}
+
+object WorldSetting {
+  def setDifficulty(worldNameList: List[String], difficulty: Difficulty): Unit = {
+    worldNameList.foreach { name =>
+      val world = Bukkit.getWorld(name)
+
+      if (world == null)
+        Bukkit.getLogger.warning(name + "という名前のワールドは存在しません。")
+      else
+        world.setDifficulty(difficulty)
+    }
+  }
+}
